Refactor with a base class
