+++ src/Umbraco.Core/Routing/ContentFinderByIdentifierPathBase.cs
@@ -1,0 +1,43 @@
+using Microsoft.Extensions.Logging;
+using Umbraco.Cms.Core.Web;
+
+namespace Umbraco.Cms.Core.Routing;
+
+public abstract class ContentFinderByIdentifierPathBase
+{
+    private readonly IRequestAccessor _requestAccessor;
+    private readonly ILogger<ContentFinderByIdentifierPathBase> _logger;
+
+    /// <remark>
+    ///     Used as the log message inside <see cref="LogAndReturnFailure()"/>>.
+    /// </remark>
+    protected abstract string FailureLogMessageTemplate { get; }
+
+    protected ContentFinderByIdentifierPathBase(IRequestAccessor requestAccessor, ILogger<ContentFinderByIdentifierPathBase> logger)
+    {
+        _requestAccessor = requestAccessor;
+        _logger = logger;
+    }
+
+    protected void ResolveAndSetCultureOnRequest(IPublishedRequestBuilder frequest)
+    {
+        var cultureFromQuerystring = _requestAccessor.GetQueryStringValue("culture");
+
+        // if we have a node, check if we have a culture in the query string
+        if (!string.IsNullOrEmpty(cultureFromQuerystring))
+        {
+            // we're assuming it will match a culture, if an invalid one is passed in, an exception will throw (there is no TryGetCultureInfo method), i think this is ok though
+            frequest.SetCulture(cultureFromQuerystring);
+        }
+    }
+
+    protected Task<bool> LogAndReturnFailure()
+    {
+        if (_logger.IsEnabled(LogLevel.Debug))
+        {
+            _logger.LogDebug(FailureLogMessageTemplate);
+        }
+
+        return Task.FromResult(false);
+    }
+}
