+++ src/Umbraco.Core/Routing/ContentFinderByKeyPath.cs
@@ -12,12 +12,13 @@ namespace Umbraco.Cms.Core.Routing;
 /// <remarks>
 ///     <para>Handles <c>/e7b65017-c6b3-4c11-b7c7-7ea1d0404c9a</c> where <c>e7b65017-c6b3-4c11-b7c7-7ea1d0404c9a</c> is the key of a document.</para>
 /// </remarks>
-public class ContentFinderByKeyPath : IContentFinder
+public class ContentFinderByKeyPath : ContentFinderByIdentifierPathBase, IContentFinder
 {
     private readonly ILogger<ContentFinderByKeyPath> _logger;
-    private readonly IRequestAccessor _requestAccessor;
     private readonly IUmbracoContextAccessor _umbracoContextAccessor;
     private WebRoutingSettings _webRoutingSettings;
+
+    protected override string FailureLogMessageTemplate => "Not a node key";
 
     /// <summary>
     ///     Initializes a new instance of the <see cref="ContentFinderByKeyPath" /> class.
@@ -27,11 +28,11 @@ public class ContentFinderByKeyPath : IContentFinder
         ILogger<ContentFinderByKeyPath> logger,
         IRequestAccessor requestAccessor,
         IUmbracoContextAccessor umbracoContextAccessor)
+        : base(requestAccessor, logger)
     {
         _webRoutingSettings = webRoutingSettings.CurrentValue ??
                               throw new ArgumentNullException(nameof(webRoutingSettings));
         _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-        _requestAccessor = requestAccessor ?? throw new ArgumentNullException(nameof(requestAccessor));
         _umbracoContextAccessor =
             umbracoContextAccessor ?? throw new ArgumentNullException(nameof(umbracoContextAccessor));
 
@@ -90,14 +91,7 @@ public class ContentFinderByKeyPath : IContentFinder
             return LogAndReturnFailure();
         }
 
-        var cultureFromQuerystring = _requestAccessor.GetQueryStringValue("culture");
-
-        // if we have a node, check if we have a culture in the query string
-        if (!string.IsNullOrEmpty(cultureFromQuerystring))
-        {
-            // we're assuming it will match a culture, if an invalid one is passed in, an exception will throw (there is no TryGetCultureInfo method), i think this is ok though
-            frequest.SetCulture(cultureFromQuerystring);
-        }
+        ResolveAndSetCultureOnRequest(frequest);
 
         frequest.SetPublishedContent(node);
         if (_logger.IsEnabled(LogLevel.Debug))
@@ -106,15 +100,5 @@ public class ContentFinderByKeyPath : IContentFinder
         }
 
         return Task.FromResult(true);
-    }
-
-    private Task<bool> LogAndReturnFailure()
-    {
-        if (_logger.IsEnabled(LogLevel.Debug))
-        {
-            _logger.LogDebug("Not a node key");
-        }
-
-        return Task.FromResult(false);
     }
 }
